add `IpClass::Auto` option
The new `IpClass::Auto` variant enables announcing on one or both ipv4
and ipv6, if the `Socket` is able to bind on that ip class. This is
different from `IpClass::V4AndV6` because that enum variant will cause
an error in `Socket::new` if the socket is not able to bind on both ipv4
and ipv6. `IpClass::Auto` will error only if the socket is not able to
bind to either ip class.